requirements.txt
flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
flask_sqlalchemy==3.1.1
